+++ src/main/java/com/github/pdfhandler4j/imp/ByOddPagesPdfSplitter.java
@@ -6,7 +6,7 @@ public class ByOddPagesPdfSplitter extends ByParityPdfSplitter {
   }
   
   @Override
-  protected String computeFileName(int beginPage) {
+  protected String computeFileName(long beginPage) {
     return "(páginas ímpares)";
   }
